Fq1jjeR

+++ Semester_1/Fq1jjeR/HW_6/3_Task.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int N, M;
+    long long N, M;
 
     cout << "Введите количество N (строк): ";
     cin >> N;
@@ -11,8 +11,8 @@ int main() {
     cout << "Введите количевство M (столбцов): ";
     cin >> M;
 
-    int count_N = N * (N + 1) / 2;
-    int count_M = M * (M + 1) / 2;
+    long long count_N = N * (N + 1) / 2;
+    long long count_M = M * (M + 1) / 2;
 
     cout << "Количество прямоугольников: " << count_N * count_M << endl;
